pyproject.toml
[tool.poetry]
name = "trtutor"
version = "0.1.0"
description = ""
authors = ["vb-creator <varshinibala@umass.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.75.0"
anthropic = "^0.49.0"
pyyaml = "^6.0.2"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
